[#429] Move TreeServiceInterface to layer back
